% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosd.R
\name{cosd}
\alias{cosd}
\title{Computes the cosine of an angle expressed in decimal degrees.}
\usage{
cosd(degrees)
}
\arguments{
\item{degrees}{either a number or a numeric vector of angles in decimal degrees.}
}
\value{
number or numeric vector.
}
\description{
Computes the cosine of an angle expressed in decimal degrees.
}
\examples{
cosd(90)
plot(seq(0,360,1), cosd(seq(0,360,1)), type='l')
}
\author{
Luca Butikofer
}

