% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{distances}
\alias{distances}
\title{Hidden functions for EM algorithm.}
\usage{
distances(dati)
}
\description{
These functions are used by the EM() function and were contained in the file 'distances.R'.
}
\author{
Beatrix Jones, Luca Butikofer
}
\keyword{internal}

