% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{frogsJK}
\alias{frogsJK}
\title{Four Jackknife re-samplings done on dataset 'frogs' generated by function jackKnife().}
\format{
A list of four dataframes with 164 rows and 6 columns each.
}
\usage{
data(frogsJK)
}
\description{
List of four Jackknife re-samplings of dataframe 'frogs' as returned by function jackKnife()(see examples in ?jackKnife).
}
\details{
\itemize{
  \item year: year of the sighting
  \item y: latitude in the New Zealand Transverse Mercatore coordinate system (length unit of measure: meter).
  \item x: longitude in the New Zealand Transverse Mercatore coordinate system (length unit of measure: meter).
  \item species: the name of the species.
}
}
\keyword{datasets}
