% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{frogsSum}
\alias{frogsSum}
\title{Summary of the dissimilarity values of each of the dataframes in 'frogsLacro' with dataframe 'frogs'.}
\format{
list of 8 lists of 10 dataframes each with 164 rows and 6 columns each.
}
\usage{
data(frogsSum)
}
\description{
Dataframe with dissimilarity values and respective Alpha value as obtained by function modSel() (see examples in ?modSel).
}
\details{
\itemize{
  \item dissimilarity: dissimilarity value
  \item compAlpha: alpha value of the comparison dataset.
}
}
\keyword{datasets}
