% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPG.R
\name{NPG}
\alias{NPG}
\title{Generates points of natural origin.}
\usage{
NPG(PresLoc, N, Deltas, year)
}
\arguments{
\item{PresLoc}{data frame with coordinates of sighting locations. Column 2 must be 'y' (latitude on a projected coordinate system with meters as distance unit of measure) and column 3 must be 'x' (longitude on a projected coordinate system with meters as distance unit of measure).}

\item{N}{total number of points to generate.}

\item{Deltas}{y and x shifts as generated from function SDS(). This function samples distance and angle of new points form existing ones from this argument.}

\item{year}{the year lable to attach to the generated points}
}
\value{
data frame of combined source ('PresLoc' argument) and new locations.
}
\description{
This function generates points of natural origin around existent points by sampling their distance from the original propagule from a user-defined dispersal kernel (as set in function SDS(), the output of which is to be entered in this function as argument 'Delta'.
}
\author{
Luca Butikofer
}

