% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAlpha.R
\name{plotAlpha}
\alias{plotAlpha}
\title{Plots the output of modSel().}
\usage{
plotAlpha(SSIM, REP, BP = FALSE)
}
\arguments{
\item{SSIM}{data frame output of modSel.}

\item{REP}{number of replicates.}

\item{BP}{when FALSE (default) plots a line representing the average of all replicates. When TRUE plots boxplots of all replicates for each alpha value.}
}
\value{
plot.
}
\description{
This function makes a plot of the data frame outputted by function modSel(). Useful to choose wich of the simulated datasets is the most similar to the observed one.
}
\examples{
data(frogsSum)
plotAlpha(SSIM= frogsSum, REP= 10)
plotAlpha(SSIM= frogsSum, REP= 10, BP=TRUE)
}
\author{
Luca Butikofer
}

