% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPG.R
\name{RPG}
\alias{RPG}
\title{Generates random points within a geographic boundary.}
\usage{
RPG(rpopn, boundary, year = "anyYear", SP)
}
\arguments{
\item{rpopn}{total number of random points to be generated.}

\item{boundary}{either a RasterLayer object (0: inside; >= 1: outside) or a SpatialPolygon(s) object.}

\item{year}{year name (used by wrapper function simulacro() to lable consecutive time steps in its simulate time series).}

\item{SP}{name of the species being simulated for labelling the output.}
}
\value{
data frame of random points.
}
\description{
This function generates random points within a geographic area set as either sp::SpatialPolygons or raster::RasterLayer.
}
\author{
Luca Butikofer
}

