% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sind.R
\name{sind}
\alias{sind}
\title{Computes the sine of an angle expressed in decimal degrees.}
\usage{
sind(degrees)
}
\arguments{
\item{degrees}{either a number or a numeric vector of angles in decimal degrees.}
}
\value{
number or numeric vector.
}
\description{
Computes the sine of an angle expressed in decimal degrees.
}
\examples{
sind(90)
plot(seq(0,360,1), sind(seq(0,360,1)),type='l')
}
\author{
Luca Butikofer
}

