% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{frogs}
\alias{frogs}
\title{Frogs sighting locations.}
\format{A dataframe with 194 rows and 4 columns}
\source{
Selection of sightings locations form the Herpetofauna Database of New Zealand. Kindly provided by Benno Kappers, Department of Conservation.
}
\usage{
data(frogs)
}
\description{
Dataset containing sighting locations of L. raniformis in New Zealand.
}
\details{
\itemize{
  \item year: year of the sighting
  \item y: latitude in the New Zealand Transverse Mercatore coordinate system (length unit of measure: meter).
  \item x: longitude in the New Zealand Transverse Mercatore coordinate system (length unit of measure: meter).
  \item species: the name of the species.
}
}
\keyword{datasets}

