% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{frogsEM}
\alias{frogsEM}
\title{Frogs sighting locations and the output of the EM() function.}
\format{A dataframe with 194 rows and 6 variables}
\usage{
data(frogsEM)
}
\description{
Version of 'frogs' Dataset containing sighting locations of L. raniformis in New Zealand as returned by function EM().
}
\details{
\itemize{
  \item year: year of the sighting
  \item y: latitude in the New Zealand Transverse Mercatore coordinate system (length unit of measure: meter).
  \item x: longitude in the New Zealand Transverse Mercatore coordinate system (length unit of measure: meter).
  \item species: the name of the species.
  \item Pnat: probability of being of natural origin [0;1].
  \item Dist: distance from nearest point of natural origin or form anchor point.
}
}
\keyword{datasets}

