% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{frogsLacro}
\alias{frogsLacro}
\title{Simulated datasets for comparison with 'frogs' generated with function simulacro() (see examples in ?simulacro()).}
\format{list of 8 lists of 10 data frames each with 164 rows and 6 columns each.}
\usage{
data(frogsLacro)
}
\description{
List of eight lists (one per different Alpha value used) each containing ten data frames of simulated sigthing locations built to emulate dataset 'frogs' (data(frogs)).
The columns of all 80 data sets are as follows:
}
\details{
\itemize{
  \item year: year of the sighting
  \item y: latitude in the New Zealand Transverse Mercatore coordinate system (length unit of measure: meter).
  \item x: longitude in the New Zealand Transverse Mercatore coordinate system (length unit of measure: meter).
  \item species: the name of the species.
  \item Pnat: probability of being of natural origin [0;1].
  \item Dist: distance from nearest point of natural origin or form anchor point.
}
}
\keyword{datasets}

