% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{frogsLacro}
\alias{frogsLacro}
\title{Simulated datasets for comparison with 'frogs' dataset generated with function simulacro() (see examples in ?simulacro).}
\format{list of 8 lists of 10 dataframes each with 164 rows and 6 columns each.}
\usage{
data(frogsLacro)
}
\description{
List of eight lists (one per different Alpha value used) each containing ten dataframes of simulated sigthing locations built to simulate 'frogs' dataset.
}
\details{
\itemize{
  \item year: year of the sighting
  \item y: latitude in the New Zealand Transverse Mercatore coordinate system (length unit of measure: meter).
  \item x: longitude in the New Zealand Transverse Mercatore coordinate system (length unit of measure: meter).
  \item species: the name of the species.
  \item Pnat: probability of being of natural origin [0;1].
  \item Dist: distance from nearest point of natural origin or form anchor point.
}
}
\keyword{datasets}

