% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx.R
\name{fx}
\alias{fx}
\title{One dimentional dispersal kernel function.}
\usage{
fx(x, a, c)
}
\arguments{
\item{x}{vector of distances.}

\item{a}{Alpha value.}

\item{c}{C value.}
}
\value{
numeric vector.
}
\description{
Computes user-defined one dimensional dispersal kernels.
}
\examples{
plot(1:1000, fx(x= 1:1000, a= 300, c=2), type='l',
 ylab= 'Probablitiy', xlab='Distance',
 main= 'One Dimensional Dispersal Kernel \\n Alpha= 300; C= 2')
}
\author{
Luca Butikofer
}

