% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{nzp}
\alias{nzp}
\title{Landmasses of the two main islands of New Zealand.}
\format{object of class sp::SpatialPolygons.}
\usage{
data(nzp)
}
\description{
Object of class sp::SpatialPolygons of the North and South Islands of New Zealand in the New Zealand Transverse Mercatore projection.
Source of data: mapdata::worldHires().
}
\keyword{datasets}

