% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{nzw}
\alias{nzw}
\title{Landmasses of the two main islands of New Zealand as Window object.}
\format{object of class spatstat::owin.}
\usage{
data(nzw)
}
\description{
Object of class spatstat::owin of the North and South Islands of New Zealand in the New Zealand Transverse Mercatore projection obtained from object 'nzp' (see ?nzp).
}
\keyword{datasets}

